Add files via upload
An individual’s annual income results from various factors. Intuitively, it is influenced by the individual’s education level, age, gender, occupation, and etc.
The dataset contains 15 columns, in which the Target Field: Income (The Income is divided into two classes: <=50K and >50K).
Other Number of Attributes: 14. These are the Demographics and other features needed to describe a person.

The Prediction Task is to determine whether a Person makes over 50K Income in a year.
Also, Objective is to perform EDA, find the missing values if any, find the outliers, and lastly build various Machine Learning models considering ‘income’ as target variable and compare the performance of each of the ML Model.