Add files via upload
Task - Open Pub Application

Let’s assume you are on a vacation in the United Kingdom with your friends. Just for some fun, you decided to go to the Pubs nearby for some drinks. Google Map is down because of some issues. 

While searching the internet, you came across https://lnkd.in/g2H6-hg9. On this website, you found all the pub locations (Specifically Latitude and Longitude info). In order to impress your friends, you decided to create a web application with the data available in your hand.